+++ laboratoires/lab04-poo3/solution/lab04b-c/Duration.h
@@ -5,23 +5,23 @@
 
 class Duration {
 private:
-    unsigned int _totalSeconds;
+    unsigned long int _totalSeconds;
 public:
     Duration() : Duration(0) {}
-    Duration(unsigned int totalSeconds);
+    Duration(unsigned long int totalSeconds);
     Duration(unsigned int hours, unsigned int minutes, unsigned int seconds);
 
-    unsigned int getTotalSeconds() const;
+    unsigned long int getTotalSeconds() const;
     unsigned int getHours() const;
     unsigned int getMinutes() const;
     unsigned int getSeconds() const;
 
     Duration& addHours(unsigned int hours);
     Duration& addMinutes(unsigned int minutes);
-    Duration& addSeconds(unsigned int seconds);
+    Duration& addSeconds(unsigned long int seconds);
     Duration& subtractHours(unsigned int hours);
     Duration& subtractMinutes(unsigned int minutes);
-    Duration& subtractSeconds(unsigned int seconds);
+    Duration& subtractSeconds(unsigned long int seconds);
 
     std::string toString() const;
